use same logic for stack unwinder, re-use functions
