+++ src/main.rs
@@ -59,6 +59,9 @@ use backend::speedscope_receiver::SpeedscopeReceiver;
 use backend::vpp_receiver::VPPReceiver;
 use backend::foc_receiver::FOCReceiver;
 use backend::vbb_receiver::VBBReceiver;
+// re-use elf-decoding and map construction
+use backend::stack_unwinder::decode_elf_instructions;
+use backend::stack_unwinder::build_insn_map;
 // error handling
 use anyhow::Result;
 // logging
@@ -190,7 +193,7 @@ fn trace_decoder(args: &Args, mut bus: Bus<Entry>) -> Result<()> {
     let elf = object::File::parse(&*elf_buffer)?;
     assert!(elf.architecture() == object::Architecture::Riscv64);
 
-    // Initialize Capstone disassembler
+    // Initialize Capstone disassembler.
     let cs = Capstone::new()
         .riscv()
         .mode(ArchMode::RiscV64)
@@ -198,44 +201,12 @@ fn trace_decoder(args: &Args, mut bus: Bus<Entry>) -> Result<()> {
         .detail(true)
         .build()?;
 
-    // First phase: decode instructions from all executable sections
-    let mut all_decoded_insns = Vec::new();
-    println!("Processing executable ELF sections:");
-    for section in elf.sections() {
-        let name = section.name().unwrap_or("<unnamed>");
-        let flags = section.flags();
+    let all_decoded_insns = decode_elf_instructions(&elf, &cs)?;
+    let insn_map = build_insn_map(&all_decoded_insns);
 
-        if let object::SectionFlags::Elf { sh_flags } = flags {
-            if sh_flags & (SHF_EXECINSTR as u64) != 0 {
-                let entry_point = section.address();
-                let text_data = section.data()?;
-                
-                println!("  Executable Section: {} at {:#x}", name, entry_point);
-
-                let decoded_instructions = cs.disasm_all(&text_data, entry_point)?;
-                debug!("[main] {}: found {} instructions", name, decoded_instructions.len());
-
-                // Save the decoded instructions for later use
-                all_decoded_insns.push(decoded_instructions);
-            }
-        }
-    }
-
-    if all_decoded_insns.is_empty() {
-        return Err(anyhow::anyhow!("No executable instructions found in ELF file"));
-    }
-
-    // Second phase: populate insn_map using references to instructions in all_decoded_insns
-    let mut insn_map: HashMap<u64, &Insn> = HashMap::new();
-    for decoded_instructions in &all_decoded_insns {
-        for insn in decoded_instructions.iter() {
-            insn_map.insert(insn.address(), insn);
-        }
-    }
-
+    // ... continue with the rest of trace_decoder logic ...
     let encoded_trace_file = File::open(args.encoded_trace.clone())?;
-    let mut encoded_trace_reader : BufReader<File> = BufReader::new(encoded_trace_file);
-
+    let mut encoded_trace_reader = BufReader::new(encoded_trace_file);
     let mut bp_counter = BpDoubleSaturatingCounter::new(args.bp_entries);
     let mut hit_count = 0;
     let mut miss_count = 0;
